Former-commit-id: ac4efab4b1e91b86c3d7d43be558894b24618390

+++ sys/dev/ath/ath_rate/sample/sample.h
@@ -307,7 +307,7 @@ name|sn
 parameter_list|,
 name|rix
 parameter_list|)
-value|(((sn)->ratemask& (1<<(rix))) != 0)
+value|(((uint64_t) (sn)->ratemask& (1ULL<<((uint64_t) rix))) != 0)
 end_define
 
 begin_ifndef
